pyproject.toml
[tool.poetry]
name = "slv-phs-tsk-web"
version = "0.1.0"
description = ""
authors = ["kotwiceal <kotvitskyaya@gmail.com>"]
readme = "README.md"
packages = [{include = "slv_phs_tsk_web"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
flask = "^2.3.2"
redis = "^4.6.0"
flask-socketio = "^5.3.5"
plotly = "^5.15.0"
scipy = "^1.11.1"
nbformat = "^5.9.2"
dill = "^0.3.7"
psycopg2 = "^2.9.7"
pandas = "^2.0.3"
pytest = "^7.4.0"
flask-login = "^0.6.2"
flask-sqlalchemy = "^3.0.5"
flask-bcrypt = "^1.0.1"
python-dotenv = "^1.0.0"
flask-session = "^0.5.0"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"